requirements.txt
pygame==2.6.1
python-chess==1.999
